Fix: Models are not editable when run from wsgi #2844
Models need to be validate to register them with admin. From runserver
it have built in validation, and if DEBUG is True then also built in
validation is there.

In usual wsgi script there is no validation for models, so on adding
validation in wsgi script models are editable now.

Review ID: 494001